+++ src/utilities.jl
@@ -20,14 +20,6 @@ is_one_based_unit_range(itr::AbstractUnitRange{T}) where {T} =
     first(itr) == oneunit(T)
 
 """
-    to_eltype(A, x)
-
-lazily yields `x` converted to the type of the elements of `A`.
-
-"""
-to_eltype(A, x) = to_type(eltype(A), x)
-
-"""
     in_range(i, len::Integer)
 
 yields whether `1 ≤ i ≤ len`.